Extract `getDataHref` to a function
