+++ packages/next/client/page-loader.ts
@@ -1,26 +1,12 @@
 import type { ComponentType } from 'react'
+import type { DataHrefParams } from './get-data-href'
 import type { RouteLoader } from './route-loader'
-import { addBasePath } from './add-base-path'
-import { interpolateAs } from '../shared/lib/router/router'
-import getAssetPathFromRoute from '../shared/lib/router/utils/get-asset-path-from-route'
-import { addLocale } from './add-locale'
-import { isDynamicRoute } from '../shared/lib/router/utils/is-dynamic'
-import { parseRelativeUrl } from '../shared/lib/router/utils/parse-relative-url'
-import { removeTrailingSlash } from '../shared/lib/router/utils/remove-trailing-slash'
+import { getDataHref } from './get-data-href'
 import {
   createRouteLoader,
   getClientBuildManifest,
   getMiddlewareManifest,
 } from './route-loader'
-
-function normalizeRoute(route: string): string {
-  if (route[0] !== '/') {
-    throw new Error(`Route name should start with a "/", got "${route}"`)
-  }
-
-  if (route === '/') return route
-  return route.replace(/\/$/, '')
-}
 
 declare global {
   interface Window {
@@ -122,51 +108,8 @@ export default class PageLoader {
     }
   }
 
-  /**
-   * @param {string} href the route href (file-system path)
-   * @param {string} asPath the URL as shown in browser (virtual path); used for dynamic routes
-   * @returns {string}
-   */
-  getDataHref({
-    href,
-    asPath,
-    ssg,
-    flight,
-    locale,
-  }: {
-    href: string
-    asPath: string
-    ssg?: boolean
-    flight?: boolean
-    locale?: string | false
-  }): string {
-    const { pathname: hrefPathname, query, search } = parseRelativeUrl(href)
-    const { pathname: asPathname } = parseRelativeUrl(asPath)
-    const route = normalizeRoute(hrefPathname)
-
-    const getHrefForSlug = (path: string) => {
-      if (flight) {
-        return path + search + (search ? `&` : '?') + '__flight__=1'
-      }
-
-      const dataRoute = getAssetPathFromRoute(
-        removeTrailingSlash(addLocale(path, locale)),
-        '.json'
-      )
-      return addBasePath(
-        `/_next/data/${this.buildId}${dataRoute}${ssg ? '' : search}`,
-        true
-      )
-    }
-
-    const isDynamic: boolean = isDynamicRoute(route)
-    const interpolatedRoute = isDynamic
-      ? interpolateAs(hrefPathname, asPathname, query).result
-      : ''
-
-    return isDynamic
-      ? interpolatedRoute && getHrefForSlug(interpolatedRoute)
-      : getHrefForSlug(route)
+  getDataHref(params: Omit<DataHrefParams, 'buildId'>): string {
+    return getDataHref({ ...params, buildId: this.buildId })
   }
 
   /**
